Refactor to fetching tasks via viewmodel
